requirements.txt
streamlit
langchain
langchain_ollama
langchain_chroma
langchain_community
llama-index
PyPDF2
faiss-cpu
chromadb
ollama
python-pptx
